immigrated is_first to gkr example
