+++ crates/examples/src/blake/preprocessed_columns.rs
@@ -1,4 +1,4 @@
-use stwo_constraint_framework::preprocessed_columns::{IsFirst, PreProcessedColumnId};
+use stwo_constraint_framework::preprocessed_columns::PreProcessedColumnId;
 use stwo_prover::core::fields::m31::BaseField;
 use stwo_prover::core::poly::circle::CanonicCoset;
 use stwo_prover::core::ColumnVec;
@@ -70,13 +70,10 @@ impl XorTable {
             })
             .collect();
 
-        let mut constant_trace = [a_col, b_col, c_col]
+        [a_col, b_col, c_col]
             .map(|x| {
                 CircleEvaluation::new(CanonicCoset::new(self.column_bits()).circle_domain(), x)
             })
-            .to_vec();
-        // TODO!(ShaharS): Remove this line.
-        constant_trace.push(IsFirst::new(self.column_bits()).gen_column_simd());
-        constant_trace
+            .to_vec()
     }
 }
